Move categories APIs into separate directory
