+++ admin-portal-frontend/src/app/all-page/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import FilterIcon from "../icons/filter.svg";
 
-import { deletePage, getAllCategories } from "../components/categoryRoutes";
+import { deletePage, getAllCategories } from "../api/Categories";
 import PageContainer from "../components/PageContainer";
 import Toast from "../components/Toast";
