+++ admin-portal-frontend/src/app/emergencies/page.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 
 import { CategoryContainer } from "../components/CategoryContainer";
 import Toast from "../components/Toast";
-import { deleteCategory, getAllCategories } from "../components/categoryRoutes";
+import { deleteCategory, getAllCategories } from "../api/Categories";
 
 export default function CategoriesPage() {
   const [categories, setCategories] = useState([]);
